+++ sampledb/logic/background_tasks/__init__.py
@@ -6,6 +6,7 @@ from .core import start_handler_threads, stop_handler_threads, post_background_t
 from .send_mail import post_send_mail_task
 from .background_dataverse_export import post_dataverse_export_task
 from .poke_components import post_poke_components_task
+from .trigger_webhooks import post_trigger_object_log_webhooks
 
 __all__ = [
     'core',
@@ -19,4 +20,5 @@ __all__ = [
     'get_background_task_result',
     'post_dataverse_export_task',
     'post_poke_components_task',
+    'post_trigger_object_log_webhooks',
 ]