+++ docs/changelog.rst
@@ -10,6 +10,8 @@ Currently in development.
 - Allow automatic calculation of quantity values
 - Show object form errors as expandable alert message
 - Added file name search
+- Added object_log API endpoint
+- Added object log webhooks
 
 Version 0.24.1
 --------------